Considered multiply advertisers and slots.

+++ Bid.py
@@ -8,7 +8,7 @@ class Bid:
 	def __init__(self, line):
 		'@param line from file from which the value and name will be read'
 		if line is not None:
-			self.value = int(line.split('\t')[0])
+			self.value = float(line.split('\t')[0])
 			self.name = line.split('\t')[1]
 		else:
 			self.value = 0.0
